requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-storage
google-cloud-logging
google-auth